Update two sentences in intro.md
The first edit adds "for" in front of each enumeration following "designed" because they are so different that the "for" helps make them consistent and clarifies that they are enumerations. The very first "for" is sandwiched between the verb-adverb "designed from the ground up" and "creating"... it could modify either one. "for creating [...], for speed, and for easier reading" is easier to understand than "for creating[...], speed, and easier reading" if the reader mistakenly associates "for" with "creating" instead of "designing". There are other ways to simplify the sentence, but this keeps the original intent IMO.

The second edit is more obvious: "For example, place me" doesn't fit with the writing style. Just below there is "For example, to nudge" so I modeled the change for the "place me" sentence after the "to nudge" sentence.